requirements.txt
flask
sounddevice
simpleaudio
scipy
TTS
openai-whisper